+++ example/frontend_test_cases/single_assignment/single_assignment.f90
@@ -1,5 +1,6 @@
 program main
     implicit none
-    real(8) :: x
+    integer(4) :: x
+
     x = 1
 end program main
